+++ packages/build-utils/src/index.ts
@@ -138,6 +138,8 @@ export {
 };
 
 export { EdgeFunction } from './edge-function';
+export { ContainerImage } from './container-image';
+export type { ContainerImageConfig } from './container-image';
 export { readConfigFile, getPackageJson } from './fs/read-config-file';
 export { normalizePath } from './fs/normalize-path';
 export { getOsRelease, getProvidedRuntime } from './os';
@@ -151,6 +153,7 @@ export {
   getMaxDurationSchema,
 } from './max-duration';
 export * from './package-manifest';
+export * from './deploy-manifest';
 export { generateProjectManifest } from './node-diagnostics';
 export {
   generateRubyProjectManifest,