+++ packages/go/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @vercel/go
 
+## 3.10.1
+
+### Patch Changes
+
+- 4b90a10: Route requests for named V2 standalone Go services to their internal function.
+
+## 3.10.0
+
+### Minor Changes
+
+- e6759d0: Add WebSocket upgrade support for standalone Go servers.
+
+## 3.9.2
+
+### Patch Changes
+
+- 2158ab6: [vc dev] Add support to `experimentalServicesV2` for bindings in `vc dev`
+
 ## 3.9.1
 
 ### Patch Changes